fix. batch 적용에 따른 디비 변경
