+++ src/main/resources/application.properties
@@ -23,6 +23,10 @@ spring.datasource.hikari.maximum-pool-size=5
 spring.jpa.properties.hibernate.jdbc.batch_size=50
 spring.jpa.properties.hibernate.order_inserts=true
 spring.jpa.properties.hibernate.order_updates=true
+
+# batch
+spring.batch.jdbc.initialize-schema=always
+
 #gemini
 langchain4j.google-ai-gemini.chat-model.api-key=${API-KEY}
 langchain4j.google-ai-gemini.chat-model.model-name=gemini-2.0-flash
